Add SchemaStringConverter
