+++ RecordBuffers/src/main/java/datamine/storage/idl/Schema.java
@@ -15,114 +15,107 @@
  */
 package datamine.storage.idl;
 
-import java.util.List;
-
-import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import datamine.storage.idl.type.FieldType;
+
+import java.util.List;
 
 /**
  * Schema definition of DataMine
- * 
+ * <p/>
  * <p>
- * The schema is the biggest concept in the DataMine schema. It is 
- * a collection of tables. 
+ * The schema is the biggest concept in the DataMine schema. It is
+ * a collection of tables.
  * </p>
- * 
+ * <p/>
  * <p>
  * A schema must have the following attributes:
- * <i> The name 
+ * <i> The name
  * <i> A list of tables
  * </p>
- * 
+ *
  * @author tliu
  * @author yqi
  */
 public class Schema implements Element {
-	
-	private final String name;
-	private final List<Table> tableList;
+
+    private final String name;
+    private final List<Table> tableList;
+
+    public Schema(String name, List<Table> tables) {
+        this.name = name;
+        this.tableList = tables;
+    }
+
+    public String getName() {
+        return this.name;
+    }
+
+    /**
+     * get table list from schema
+     *
+     * @return ArrayList of tables
+     */
+    public List<Table> getTableList() {
+        return tableList;
+    }
+
+    @Override
+    public String toString() {
+        return new GsonBuilder().setPrettyPrinting().create().toJson(this);
+    }
+
+    @Override
+    public void accept(ElementVisitor visitor) {
+        visitor.visit(this);
+        for (Table table : tableList) {
+            table.accept(visitor);
+        }
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((name == null) ? 0 : name.hashCode());
+        result = prime * result
+                + ((tableList == null) ? 0 : tableList.hashCode());
+        return result;
+    }
 
 
-	public static Schema getSchema(String string){
-		Gson gson = new GsonBuilder().registerTypeAdapter(FieldType.class, new Field.FieldDeserializer()).create();
-		return gson.fromJson(string, Schema.class);
-	}
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
 
-	public Schema(String name, List<Table> tables) {
-		this.name = name;
-		this.tableList = tables;
-	}
-	
-	public String getName() {
-		return this.name;
-	}
-	
-	/**
-	 * get table list from schema
-	 * @return ArrayList of tables
-	 */
-	public List<Table> getTableList() {
-		return tableList;
-	}
+        if (obj == null) {
+            return false;
+        }
 
-	@Override
-	public String toString() {
-		return new GsonBuilder().setPrettyPrinting().create().toJson(this);
-	}
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
 
-	@Override
-	public void accept(ElementVisitor visitor) {
-		visitor.visit(this);
-		for (Table table : tableList) {
-			table.accept(visitor);
-		}
-	}
+        Schema other = (Schema) obj;
+        if (name == null) {
+            if (other.name != null) {
+                return false;
+            }
 
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((name == null) ? 0 : name.hashCode());
-		result = prime * result
-				+ ((tableList == null) ? 0 : tableList.hashCode());
-		return result;
-	}
+        } else if (!name.equals(other.name)) {
+            return false;
+        }
 
+        if (tableList == null) {
+            if (other.tableList != null) {
+                return false;
+            }
 
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj){
-			return true;
-		}
-			
-		if (obj == null) {
-			return false;
-		}
-			
-		if (getClass() != obj.getClass()) {
-			return false;
-		}
-			
-		Schema other = (Schema) obj;
-		if (name == null) {
-			if (other.name != null) {
-				return false;
-			}
-				
-		} else if (!name.equals(other.name)) {
-			return false;
-		}
-			
-		if (tableList == null) {
-			if (other.tableList != null) {
-				return false;
-			}
-				
-		} else if (!tableList.equals(other.tableList)) {
-			return false;
-		}
-			
-		return true;
-	}
+        } else if (!tableList.equals(other.tableList)) {
+            return false;
+        }
+
+        return true;
+    }
 }
